refactor: AuthContext 분리
