+++ craco.config.js
@@ -6,6 +6,7 @@ module.exports = {
       '@components': path.resolve(__dirname, 'src/components'),
       '@styles': path.resolve(__dirname, 'src/styles'),
       '@pages': path.resolve(__dirname, 'src/pages'),
+      '@contexts': path.resolve(__dirname, 'src/contexts'),
     },
   },
 };
